requirements.txt
Flask==2.3.2
pycryptodome==3.19.0
Pillow==10.0.1
